Refactoring names to address Grace's comments
