+++ step-capstone/package.json
@@ -12,8 +12,8 @@
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "date-fns": "^2.14.0",
-    "firebase": "^7.16.1",
-    "firebaseui": "^4.6.0",
+    "firebase": "^7.17.1",
+    "firebaseui": "^4.6.1",
     "lodash": "^4.17.19",
     "material-ui-dropzone": "^3.3.1",
     "react": "^16.13.1",
